app/requirements.txt
django
djangorestframework
drf-yasg
django-filter
django-cors-headers
psycopg2-binary
gunicorn
djangorestframework-simplejwt